refactor to separe the modified probability algo in to more readavble sections starting with the ship probability
